fix(dev): seed household memberships for fixture users
